pyproject.toml
[project]
name = "pdf-bookmarkit-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.82.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "typer>=0.24.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bookmarkit"]

[project.scripts]
bookmarkit = "bookmarkit.main:app"
